requirements.txt
APScheduler==3.6.3
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
Django==3.1.7
django-cors-headers==3.7.0
djangorestframework==3.12.4
idna==2.10
pkg-resources==0.0.0
python-telegram-bot==13.4.1
pytz==2021.1
pyzmq==22.0.3
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
tornado==6.1
tzlocal==2.1
urllib3==1.26.4
zmq==0.0.0
